# openapi/foodize.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Get My Notifications",
        "operationId": "get_my_notifications_api_v1_notifications_get",
        "security": [
          {
            "OAuth2PasswordBearerWithCookie": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Notifications"
        ],
        "summary": "Delete All Notifications",
        "operationId": "delete_all_notifications_api_v1_notifications_delete",
        "security": [
          {
            "OAuth2PasswordBearerWithCookie": []
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        }
      }
    },
    "/api/v1/notifications/{notification_id}/read": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Read Notification",
        "operationId": "read_notification_api_v1_notifications__notification_id__read_post",
        "security": [
          {
            "OAuth2PasswordBearerWithCookie": []
          }
        ],
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/read-all": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Read All Notifications",
        "operationId": "read_all_notifications_api_v1_notifications_read_all_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Read All Notifications Api V1 Notifications Read All Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearerWithCookie": []
          }
        ]
      }
    },
    "/api/v1/notifications/{notification_id}": {
      "delete": {
        "tags": [
          "Notifications"
        ],
        "summary": "Delete Notification",
        "operationId": "delete_notification_api_v1_notifications__notification_id__delete",
        "security": [
          {
            "OAuth2PasswordBearerWithCookie": []
          }
        ],
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NotificationListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "unread_count": {
            "type": "integer",
            "title": "Unread Count"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "unread_count"
        ],
        "title": "NotificationListResponse"
      },
      "NotificationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "$ref": "#/components/schemas/NotificationType"
          },
          "is_read": {
            "type": "boolean",
            "title": "Is Read"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "title",
          "message",
          "type",
          "is_read",
          "created_at"
        ],
        "title": "NotificationResponse"
      },
      "NotificationType": {
        "type": "string",
        "enum": [
          "ORDER_STATUS",
          "SYSTEM"
        ],
        "title": "NotificationType"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearerWithCookie": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        }
      }
    }
  }
}